Add files via upload
Same database as the original except Chapter_Name field has been renamed to Book_Name field. Additionally, a Book_Id field has been added with Genesis = 1 and Revelation = 66. That way, the books of the Bible can be sorted in the order they appear in the Bible rather than relying on the physical order the items are entered into the database.